Start on a custom memory allocator that preallocates a pool based on profiles
